Обновление дизайна и функционал свободного времени. Удаление промежутков свободного времени, input’ы в стиле bootstrap, error-handling, название дней недели

+++ models/Freetime.php
@@ -27,20 +27,29 @@
 				return false;
 			}
 			
-			// Удаляем предыдущие данные 
-			// @todo: проверять есть ли данные. если их не существует – добавлять по отдельности
-/*
-			self::deleteAll([
-				"condition"	=> "id_student=$id_student",
-			]);
-*/
-			
-			
 			// Создаем интервалы свободного времени
 			foreach ($freetime_data as $one_day_freetime_data) {
-				$FreetimeData = new self($one_day_freetime_data);
-				$FreetimeData->id_student = $id_student;
-				$FreetimeData->save();
+				// Если промежуток уже существует и его надо удалить
+				if ($one_day_freetime_data["id"] && $one_day_freetime_data["deleted"]) {
+					// создаем массив из ID промежутков, которые нужно удалить
+					$ids_to_delete[] = $one_day_freetime_data["id"];	
+				} else {
+					// иначе если промежуток новый (и он не удален)
+					if (!$one_day_freetime_data["deleted"]) {
+						// добавляем промежуток свободного времени
+						$FreetimeData = new self($one_day_freetime_data);
+						$FreetimeData->id_student = $id_student;
+						$FreetimeData->save();	
+					}
+				}
+			}
+			
+			// Если есть что удалять
+			if ($ids_to_delete) {
+				// Удаляем
+				self::deleteAll([
+					"condition"	=> "id IN (". implode(",", $ids_to_delete) .")",
+				]);
 			}
 		}
